split bag content to its own file
